Move interface-utils to local-queries/webview
