+++ extensions/ql-vscode/src/local-queries/index.ts
@@ -2,3 +2,4 @@ export * from "./local-queries";
 export * from "./local-query-run";
 export * from "./quick-eval-code-lens-provider";
 export * from "./quick-query";
+export { WebviewReveal } from "./webview";
